fix pydantic subclass problem by refactoring some django-models bases
